Architectural audit and further refactoring, w bugfixes

+++ firmware/chuck_firmware/chuck_firmware.ino
@@ -90,13 +90,13 @@ struct __attribute__((packed)) ManualControlPacket {
     float x_axisStatus;         // X axis status
     float y_axisStatus;         // Y axis status
     float z_axisStatus;         // Z axis status
-    int x_stepSize;           // X step size
-    int y_stepSize;           // Y step size
-    int z_stepSize;           // Z step size
-    int dpad_LR;              // D-pad left-right value
-    int dpad_UD;              // D-pad up-down value
-    int bumpers;              // Bumper combined value
-    int manual_jog_speed;     // Manual jog speed
+    float x_stepSize;           // X step size
+    float y_stepSize;           // Y step size
+    float z_stepSize;           // Z step size
+    float dpad_LR;              // D-pad left-right value
+    float dpad_UD;              // D-pad up-down value
+    float bumpers;              // Bumper combined value
+    float manual_jog_speed;     // Manual jog speed
 };
 
 const size_t BINARY_PACKET_SIZE = sizeof(ManualControlPacket);
